Use parseFloat instead of parseInt in TreeMap

+++ frontend/src/components/charts/TreeMap.tsx
@@ -40,7 +40,7 @@ export const TreeMap = ({
           color: randomColor({
             luminosity: settings.darkMode ? "light" : "dark",
           }),
-          size: parseInt(row[columnIndex]),
+          size: parseFloat(row[columnIndex]),
         };
       }),
     };
